third_party/WebKit/LayoutTests/fast/viewport/viewport-legacy-merge-quirk-2-expected.txt

ALERT: viewport size 500x550 scale 2 with limits [2, 2] and userScalable false
